Split ArrayReaderBuilder into its own module (#1483) (#1485)
* Split ArrayReaderBuilder into its own module (#1483)

* Add license header
